pyproject.toml
[project]
name = "prooza"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.12.0",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.74",
    "langchain-experimental>=0.3.4",
    "langchain-huggingface>=0.3.1",
    "langchain-ollama>=0.3.6",
    "langchain-text-splitters>=0.3.9",
    "numpy>=2.3.2",
    "ollama>=0.5.3",
    "pandas>=2.3.2",
    "pypdf>=6.0.0",
    "sentence-transformers>=5.1.0",
    "streamlit>=1.48.1",
]
